Leitura dos dados para a tabela no front funcionando

+++ script.sql
@@ -7,7 +7,7 @@ create table devedores (
 	dt_nascimento date not null,
 	endereco varchar(255) null,
 	descricao varchar(255) not null,
-	valor float not null,
+	valor decimal(20,2) not null,
 	dt_vencimento date not null,
 	created_at datetime,
 	updated_at datetime
